Move data tests into separate module
